first untested implementation of public key sorting
When generating a p2sh address, the address is different if the public keys are
in a different order. My proposed solution to this to sort the public keys by
treating them as big endian numbers. I will advocate for this standard to be
adopted in the bitcoin world for all p2sh address generation. If the community
adopts a different standard, we will adopt whichever one is chosen by the
community. But for now it's this.